Moved Firebase authentication to TodoListService
