+++ app/src/main/java/com/example/todolist/MainActivity.kt
@@ -1,20 +1,14 @@
 package com.example.todolist
 
-import android.content.BroadcastReceiver
 import android.content.Intent
-import android.content.IntentFilter
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.util.Log
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.fragment.findNavController
 import androidx.navigation.ui.setupActionBarWithNavController
 import com.example.todolist.databinding.ActivityMainBinding
 import com.example.todolist.services.TodoListService
-import com.google.firebase.auth.FirebaseAuth
-import com.google.firebase.auth.ktx.auth
-import com.google.firebase.ktx.Firebase
 import kotlinx.android.synthetic.main.activity_main.*
 
 class MainActivity : AppCompatActivity() {
@@ -25,9 +19,6 @@ class MainActivity : AppCompatActivity() {
 
     private lateinit var navController: NavController
 
-    private lateinit var auth: FirebaseAuth
-
-    private lateinit var br: BroadcastReceiver
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -41,9 +32,6 @@ class MainActivity : AppCompatActivity() {
         setSupportActionBar(toolbar)
         setupActionBarWithNavController(navController)
 
-        auth = Firebase.auth
-        signInAnonymously()
-
 
     }
 
@@ -51,22 +39,9 @@ class MainActivity : AppCompatActivity() {
         return navController.navigateUp() || super.onSupportNavigateUp()
     }
 
-    override fun onResume() {
-        super.onResume()
-        registerReceiver(br, IntentFilter("com.example.todolist.REGISTER"))
-    }
-
     override fun onDestroy() {
         super.onDestroy()
         stopService()
-    }
-
-    private fun signInAnonymously() {
-        auth.signInAnonymously().addOnSuccessListener {
-            Log.d(TAG, "Login successfully ${it.user.toString()}")
-        }.addOnFailureListener {
-            Log.e(TAG, "Login failed", it)
-        }
     }
 
     fun stopService() {
