+++ app/src/main/java/com/example/todolist/services/TodoListService.kt
@@ -9,6 +9,9 @@ import android.widget.Toast
 import androidx.core.net.toUri
 import com.example.todolist.MainActivity
 import com.example.todolist.data.TodoList
+import com.google.firebase.auth.FirebaseAuth
+import com.google.firebase.auth.ktx.auth
+import com.google.firebase.ktx.Firebase
 import com.google.firebase.storage.FirebaseStorage
 import java.io.File
 import java.io.FileOutputStream
@@ -16,16 +19,20 @@ import java.io.FileOutputStream
 
 class TodoListService : Service() {
 
-    lateinit var todoList: MutableList<TodoList>
-    var onSave:((file: Uri) -> Unit)? = null
+    val TAG: String = "TodoList:TodoListService"
 
-    val TAG:String = "TodoList:TodoListService"
+    lateinit var todoList: MutableList<TodoList>
+    var onSave: ((file: Uri) -> Unit)? = null
+
+    private lateinit var auth: FirebaseAuth
 
     init {
         Log.d(TAG, "TodoListService is running...")
+        auth = Firebase.auth
+        signInAnonymously()
     }
 
-    override fun onBind(intent: Intent?): IBinder?  = null
+    override fun onBind(intent: Intent?): IBinder? = null
 
     override fun onStartCommand(intent: Intent?, flags: Int, startId: Int): Int {
         val todoList = intent?.getStringExtra("EXTRA_DATA")
@@ -42,30 +49,37 @@ class TodoListService : Service() {
         Toast.makeText(this, "Service destroyed", Toast.LENGTH_LONG).show()
     }
 
-    private fun onSave(fileName:String,content:String){
+    private fun onSave(fileName: String, content: String) {
         val path = this.getExternalFilesDir(null)
-        if(path != null){
-            val file = File(path,fileName)
-            FileOutputStream(file,true).bufferedWriter().use { writer ->
+        if (path != null) {
+            val file = File(path, fileName)
+            FileOutputStream(file, true).bufferedWriter().use { writer ->
                 writer.write(content)
             }
             this.onSave?.invoke(file.toUri())
             upload(file.toUri())
-        }
-        else {
+        } else {
             Toast.makeText(this, "Couldn't save to file", Toast.LENGTH_LONG).show()
             Log.e(TAG, "Filepath not found")
         }
     }
 
-    private fun upload(file:Uri){
+    private fun upload(file: Uri) {
         val ref = FirebaseStorage.getInstance().reference.child("Todo/${file.lastPathSegment}")
         val uploadTask = ref.putFile(file)
 
         uploadTask.addOnSuccessListener {
             Log.d(TAG, "Saved file to FireBase: ${it.toString()}")
-        }.addOnFailureListener{
+        }.addOnFailureListener {
             Log.e(TAG, "Error saving file to FireBase", it)
+        }
+    }
+
+    private fun signInAnonymously() {
+        auth.signInAnonymously().addOnSuccessListener {
+            Log.d(TAG, "Login successfully ${it.user.toString()}")
+        }.addOnFailureListener {
+            Log.e(TAG, "Login failed", it)
         }
     }
